Feat: alarm component + icon = alarmset
