+++ src/app/(main)/mypage/page.module.scss
@@ -19,14 +19,3 @@
     @include text-20-b;
   }
 }
-
-.button {
-  position: relative;
-  margin-left: 40rem;
-}
-
-.alarms {
-  position: absolute;
-  top: 3.4rem;
-  left: -34.4rem;
-}
