+++ src/components/common/AlarmSet.module.scss
@@ -1,0 +1,21 @@
+@import '../../styles/index';
+
+.button {
+  position: relative;
+  margin-left: 40rem;
+}
+
+.alarms {
+  position: absolute;
+  top: 3.4rem;
+  left: -34.4rem;
+
+  @include media(mobile) {
+    top: -4rem;
+    left: -41.5rem;
+  }
+}
+
+.closeAlarm {
+  display: none;
+}
